ship/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-storage-blob==12.24.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
configparser==7.1.0
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
openai==1.65.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
